+++ examples/H2_Analysis/h2_setup_optimize.py
@@ -13,7 +13,7 @@ from hopp.sites import SiteInfo
 from hopp.sites import flatirons_site as sample_site
 from hopp.hybrid_simulation import HybridSimulation
 from hopp.log import hybrid_logger as logger
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utilities.keys import set_developer_nrel_gov_key
 from simple_dispatch import SimpleDispatch
 from run_h2_PEM import run_h2_PEM
 import numpy as np
