+++ examples/optimization/layout_opt/hybrid_run.py
@@ -37,7 +37,7 @@ from tools.optimization import (
 from hopp.sites import make_circular_site, make_irregular_site, SiteInfo
 from hopp.log import opt_logger as logger
 from hopp.sites import locations
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utilities.keys import set_developer_nrel_gov_key
 from hopp.layout.plot_tools import *
 
 from parametrized_optimization_driver import ParametrizedOptimizationDriver
