+++ examples/optimization/layout_opt/wind_run.py
@@ -29,7 +29,7 @@ from tools.optimization import (
     DataRecorder
     )
 from hopp.sites import SiteInfo, flatirons_site
-from hopp.keys import set_developer_nrel_gov_key
+from hopp.utilities.keys import set_developer_nrel_gov_key
 
 from parametrized_optimization_driver import ParametrizedOptimizationDriver
 from wind_optimization_problem import WindOptimizationProblem
