+++ hopp/eco/hopp_mgmt.py
@@ -1,5 +1,5 @@
-from hopp.sites import SiteInfo
-from hopp.sites import flatirons_site as sample_site
+from hopp.simulation.technologies.sites import SiteInfo
+from hopp.simulation.technologies.sites import flatirons_site as sample_site
 
 from examples.H2_Analysis.hopp_for_h2_floris import (
     hopp_for_h2_floris as hopp_for_h2,
