+++ hopp/tools/hopp_tools.py
@@ -467,7 +467,7 @@ def compressor_model():
 def pressure_vessel():
 
     #Pressure Vessel Model Example
-    from hopp.hydrogen.h2_storage.pipe_storage.underground_pipe_storage import Underground_Pipe_Storage
+    from hopp.simulation.technologies.hydrogen.h2_storage.pipe_storage.underground_pipe_storage import Underground_Pipe_Storage
     storage_input = dict()
     storage_input['H2_storage_kg'] = 18750
     # storage_input['storage_duration_hrs'] = 4
